protects cleanup_heredocs() function from malloc() failures

+++ utils/memory.c
@@ -6,64 +6,13 @@
 /*   By: tpinarli <tpinarli@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/11 10:30:48 by tpinarli          #+#    #+#             */
-/*   Updated: 2025/05/23 20:21:29 by tpinarli         ###   ########.fr       */
+/*   Updated: 2025/06/01 12:42:16 by ykadosh          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
 static void	free_tokens(t_token **tokens);
-
-/*
-void	free_2D_char(char **arr)
-{
-	int i;
-
-	i = 0;
-	while (arr[i])
-		free(arr[i++]);
-	free(arr);
-}
-
-void	ft_free_split(char **arr)
-{
-	int	i;
-
-	i = 0;
-	if (!arr)
-		return ;
-	while (arr[i])
-	{
-		free(arr[i]);
-		i++;
-	}
-	free(arr);
-}
-
-void	free_env(char **env)
-{
-	int i = 0;
-	while (env[i])
-		free(env[i++]);
-	free(env);
-}
-
-
-void	free_argv(char **argv)
-{
-	int	i;
-
-	i = 0;
-	if (!argv)
-		return ;
-	while (argv[i])
-	{
-		free(argv[i]);
-		i++;
-	}
-	free(argv);
-}
-*/
 
 void	free_two_dimensional_array(char ***arr)
 {
@@ -82,21 +31,6 @@ void	free_two_dimensional_array(char ***arr)
 	free(*arr);
 	*arr = NULL;
 }
-
-/*
-static void	free_redir_list(t_redir *redir)
-{
-	t_redir	*next;
-
-	while (redir)
-	{
-		next = redir->next;
-		free(redir->filename);
-		free(redir);
-		redir = next;
-	}
-}
-*/
 
 static void	free_redir_list(t_redir **redir)
 {
@@ -118,25 +52,6 @@ static void	free_redir_list(t_redir **redir)
 	}
 }
 
-/*
- * WARN:  delete this when ready.
-void	free_cmd(t_command *cmd)
-{
-	t_command	*next;
-
-	while (cmd)
-	{
-		next = cmd->next;
-		free_argv(cmd->argv);
-		free_redir_list(cmd->in_redir);
-		free_redir_list(cmd->out_redir);
-		free(cmd);
-		cmd = next;
-	}
-}
-*/
-
-//WARN: alternate version that sets the original cmd pointer to NULL.
 void	free_cmd(t_command **cmd)
 {
 	t_command	*next;
@@ -172,22 +87,16 @@ static void	free_tokens(t_token **tokens)
 	*tokens = NULL;
 }
 
-// WARN: MAKE 100% sure that you are passing a double pointer,
-// freeing the memory properly and resetting to NULL both tokens and input!!!!
-// Same for 'cmd'!!!
-// UPDATE: this is done, but we HAVE to test it once the project compiles.
-// never pass NULL to this function! The pointer can be a NULL pointer, and
-// then you pass its address - that's OK, but never pass simply "NULL"
 void	free_all(char **input, t_token **tokens, t_command **cmd)
 {
-	if (*input)
+	if (input && *input)
 	{
 		free(*input);
 		*input = NULL;
 	}
-	if (*tokens)
+	if (tokens && *tokens)
 		free_tokens(tokens);
-	if (*cmd)
+	if (cmd && *cmd)
 	{
 		cleanup_heredocs(*cmd);
 		free_cmd(cmd);
@@ -222,9 +131,12 @@ void	free_deprecated_strings(t_token *tokens, size_t k)
 	}
 }
 
-// WARN: ft_itoa() and ft_strjoin() are unprotected here.
-// Can we free the heredocs differently? This is on the dangerous side...
-void	cleanup_heredocs(t_command *cmd)
+/*
+* return values:
+* 0: if malloc() has failed during call to ft_itoa() or ft_strjoin()
+* 1: otherwise
+*/
+int	cleanup_heredocs(t_command *cmd)
 {
 	t_redir *in;
 	char	*heredoc_file_name;
@@ -237,13 +149,18 @@ void	cleanup_heredocs(t_command *cmd)
 	{
 		if (in->type == REDIR_HEREDOC)
 		{
-			heredoc_number = ft_itoa(i); // WARN: unprotected
-			heredoc_file_name = ft_strjoin("heredoc_", heredoc_number); // WARN: unprotected
+			heredoc_number = ft_itoa(i);
+			if (!heredoc_number)
+				return (0);
+			heredoc_file_name = ft_strjoin("heredoc_", heredoc_number);
+			free(heredoc_number);
+			if (!heredoc_file_name)
+				return (0);
 			unlink(heredoc_file_name);
 			free(heredoc_file_name);
-			free(heredoc_number);
 			i++;
 		}
 		in = in->next;
 	}
+	return (1);
 }
